Added newline parameter to file writing
Added newline='' to when writing the file occurs which removes the issue of creating additional, blank lines in some systems which will in turn lead to errors when reading the file to download/crawl the same URL as it in `get_handled_list` k is > 0, but the rest of the line is empty so throws an error. This has fixed this issue.